Allow reloading images in LLNL import tool
Previously when users changed the detector or image plate that had already been
loaded and/or cropped the text was not updated and the image was not re-loaded.
Now the previously selected image is loaded, the text label is updated, and if
it is an image plate the completed template is drawn in the preciously selected
position.

Signed-off-by: Brianna Major <brianna.major@kitware.com>