Use defaults for state files with missing attributes
Previously any state files that were missing attributes would simply keep the
current settings when a new state file is loaded, but loading a state file
should be a reset of sorts. Now we apply default values when a state is loaded
if there is an attribute in attributes_to_persist but not in the state file.

Signed-off-by: Brianna Major <brianna.major@kitware.com>