Keep valid images after detectors added/removed
Previously, if detectors were added/removed from the instrument
form view, we would automatically clear all the imageseries. We can
instead, however, keep the imageseries that are still valid. If one
particular detector does not contain good images, this will allow,
for example, that detector to just be removed.

The imageseries that are considered "valid" and are kept are only
ones whose detector name still exists in the list of detectors, and
whose shape matches the detector shape.

Depends on: https://github.com/HEXRD/hexrd/pull/790

Fixes: #1821

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>