Before modifying, make copy of image if needed
For some imageseries types (particularly HDF5), the returned numpy
array is not writeable for some reason. An error is thrown if we try
to modify it in place.

So we should check if the image is writeable before trying to modify
it. If it's not writeable, we make a copy that *is* writeable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>