Fix issue where "aggregate" is disabled
This issue would happen for composite instruments where the same file is
used by multiple subpanels.

Basically, we were miscounting when we did the calculation to determine
whether aggregation should be enabled. For example, for 32 subpanel Eiger,
if there were less than 64 frames in the data, aggregation would be disabled.
The new counting should fix that.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>