Provide chemical formula for absorption correction
For all of the layers (both those on the detector and in the physics
package), provide the chemical formula, if possible, to be used for
absorption correction.

This enables users to select materials whose name is not a chemical
formula, such as "diamond". We extract the chemical formula from the
material and provide that.

Depends on: hexrd/hexrd#800

Fixes: #1824

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>