Fix panel buffer colormap range to be 0 to 1
Normally, the panel buffer colormap shows masked pixels as purple
and unmasked pixels as yellow.

However, if all pixels are unmasked, the panel buffer colormap used to
surprisingly show all pixels as purple (the opposite of what you would
expect). This is because we were allowing matplotlib to set the vmin
and the vmax automatically based upon the contents of the data. If all
of the data were all zeros, or a mixture of zeros and ones, it would
appear as we expected. However, if the data were all ones, it would
be scaled differently.

Setting the vmin and the vmax manually fixes this issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>