Add ability to select detectors to use for lineout
A new button is added in the polar view menu options named
"Select Detectors for Lineout". This allows the user to select
a subset of detectors to be used when calculating the lineout.

It works by keeping track of which pixels in the polar view are
"affected" by each detector. If a pixel is affected by a detector,
and it is not selected for the azimuthal lineout plot, that pixel
gets set to nan when computing the azimuthal lineout.

The one caveat of this implementation: if an instrument has overlapping
detectors in polar space, and a pixel is affected by both a selected
detector and an unselected detector, that pixel will still be set to nan,
despite being part of the pixels of the selected detector. This scenario
should be very rare, however, and it saves us a bunch of developer time
and computational time to do it the current way.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>