Memoize projection to polar view
This function was memoized in the past, and appears to have been
"un-memoized" by accident. After the polar view has been generated
one time, this memoization substantially improves the performance of
re-generating the same polar view using the same settings (for example,
switching to a different image mode and then coming back).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>