Keep all eta omega map HKLs in workflow export
We previously removed all HKLs other than the seed HKLs when exporting
the full HEDM workflow. We thought that the non-seed HKLs were not used.
However, they were indeed used when measuring completeness in `paintGrid()`.

This typically would only cause minor issues (a difference in the number of
orientations kept). However, if only one seed HKL was selected, the
completeness measurement would be very off, and many bad orientations would
be kept. This was because a single eta omega map is not enough to accurately
determine the completeness of an orientation - you typically need at least two
or more.

Now we keep all eta omega map HKLs when exporting the full workflow. We seem
to get nearly identical results in the GUI and the CLI with this.

Fixes: hexrd/hexrd#802

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>