Remove min subtraction in scaling
The min subtraction was causing some inconsistencies in coloring within
the same dataset, because if someone, for example, masks out (or performs
something like an erosion on) the data which eliminates the old minimum
in the data, it would appear that all of the data values are rescaled.

We also should include a `+ 1` for the square root cases that will
eliminate most negative values we'd encounter.

The only case where the data may be adversely affected by this change
would be cases where there are some very negative values in the data.
Perhaps this happens for some datasets with dark background subtraction.

Fixes: #1878

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>