Expand polar masks beyond detector border
When polar mask coordinates are right along the border, they sometimes
don't fully mask pixels along the border.

This commit shifts any coordinates that are very close to the border
to being half a pixel beyond the border. This should hopefully ensure
that the pixels are masked in the polar view.

Fixes: #1879

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>