Add beam energy correction to grains
This allows a slope (in eV/mm), intercept (in mm), and axis ('x' or 'y')
to be specified for a beam energy correction according to grain positions.

This is important to take into consideration gradients in the beam energy
profile, such that some grains may "experience" a different beam energy
than others.

The option to add the HEDM beam energy correction was added to the
instrument form view. If added, it also goes into the tree view, and the
slope and the intercept will appear in the HEDM calibration as well as
refinable parameters.

Depends on: hexrd/hexrd#819

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>