Allow no picks on subpanels in HEDM calibration
Previously, there would be an error if a detector did not have any picks
in the HEDM calibration. This would be really common for multi-subpanel
Eiger, where some of the subpanels, especially in the corners, would often
not have picks.

The calibration now works fine with no picks on the subpanels.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>