Fix tth distortion for composite calibration
Before, we were accidentally forcing the composite calibration to
use the tth_distortion settings in the calibration dialog for all
of its overlays. We don't want to do that, because each overlay
needs to have its own calibration settings respected.

Fix the code so that the overlays use their own tth_distortion when
calibrating. Also fix rerender updates for when the polar view tth
distortion settings are modified.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>