Add ability to display WPPF difference plot
This adds the ability to display a difference plot between the WPPF
curve and the azimuthal lineout as a separate plot at the bottom of
the polar view canvas.

The difference plot can be displayed as either absolute difference (default)
or a percent difference.

The difference plot is also exported with the WPPF data when an export
is performed.

Fixes: #852

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>