Disable expression parameters in WPPF
This is intended for disabling the final phase fraction parameter,
which gets set automatically so that all phase fractions sum to 1.

We also had to fix an issue where the background parameters for
chebyshev did not start out with correct values. This issue came
about due to changes needed in hexrd. That is now fixed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>