Fix powder range highlighting
Oftentimes, when a powder ring is highlighted, the powder range would
be highlighted incorrectly (typically a range on a different ring to
the left of the highlighted ring).

This was due to the fact that `rbnd_indices` are indices into the
two theta list, and not into the ring array.

We added a ring map to fix this, and now the correct range is always
highlighted.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>