Disable zoom canvas during highlight updates
They both use blitting, and unfortunately the zoom canvas does not
utilize the canvas's blit manager (which it should).

Because of that, when a blit update is being performed by both the
main canvas and the zoom canvas, there is some bouncing back and
forth in the rendering. This mainly manifests itself by the highlighted
line bouncing around.

This commit fixes the issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>