Consider tth distortion when saving picks
We previously would not consider the tth distortion when saving
picks, so the cartesian coordinates would be incorrect.

Now we reverse the tth distortion if needed. Also, the two calibration
workflows (Composite and Fast Powder) write the coordinates directly
from the overlays, which is more accurate (some precision is lost when
converting back and forth between cartesian and polar when the tth
distortion is applied, because it is a field correction).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>