Do not allow memoized median output to be modified
Make a copy of the memoized median filter output so that we can be
sure the array will not be modified.

This was causing a bug where, if you had one state file loaded in
with the median filter applied, and you load in another state file
that has the median filter applied, the memoized output was somehow
modified and made the image data invalid.

Making a copy of the memoized output ensures that it won't be modified.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>