Specify sample tilt in physics package dialog
This adds the ability to specify the sample tilt within the physics package
dialog. This sample tilt is synchronized with the sample tilt settings within
the Laue overlay editor and the Const Chi editor.

We will be using this sample tilt (convertable to sample rmat or sample normal)
in a variety of places, including WPPF and the absorption correction.

Fixes: #1801

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>