Add units and conversion capabilities to WPPF
This adds the basic capabilities for units and unit conversion to the
WPPF module, and a few initial examples, including for lattice parameters.

The lattice parameter lengths are automatically converted from nanometers
to Angstroms for display (and back when the user modifies them).

It also adds Angstrom units to the lattice lengths, and degree units to
the lattice angles.

We'll use this same infrastructure for other unit conversions.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>