Change all artist removals to use remove_artist()
It's not safe to remove artists the old way, because they may raise
an exception if the figure they were in was cleared.

Instead, we have to do it this way and just ignore any errors that
occur. It would be nice if we figured out a cleaner way at some point.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>